clients part of the programm is working
TODO -add support for rooms
     -add support for chekins(booking)
     -add support for admin system

+++ src/main/java/com/example/hotel_jpa/repositories/IClientRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 //import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface IClientRepository extends JpaRepository<Client, Integer> {
+public interface IClientRepository extends JpaRepository<Client, Long> {
     @Query("select c from Client c where c.lastName = :name")
     Client findByLastName(@Param("name") String lastName);
